docs: v4.10.4 — subagent instructions, version badges, and deep review completion state finalized
code-reviewer.md: Added 5 new Deep Review checklist categories (Base Class Correctness,
Column Type Integrity, Index Filter Correctness, Navigation Property Mapping, BranchId FK
Integrity) + 19 new Features to Fix By Default (#49-#67) covering all 42+ structural findings.

implement-agent.md: Updated Phase Awareness to v4.10.4 with deep review completion stats.
Added 18 new Features to Fix By Default (#43-#60) covering base class, column type, index,
nav property, and config patterns discovered in deep review.

README.md: Version bumped to v4.10.4 across all badges and tagline. Added deep review
completion feature description: 42+ fixes, 0 errors, 1,574 tests passing.

All three instruction files now frozen with complete deep review guidance for future
code generation and review.

+++ SalesSystem/SalesSystem.Api/Controllers/AccountsController.cs
@@ -104,7 +104,7 @@ public class AccountsController : ControllerBase
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> GetMappings([FromQuery] int? branchId, CancellationToken ct)
+    public async Task<IActionResult> GetMappings([FromQuery] short? branchId, CancellationToken ct)
     {
         var result = await _systemAccountService.GetAllMappingsAsync(branchId, ct);
         if (!result.IsSuccess)
