+++ SalesSystem/SalesSystem.Application/Interfaces/Services/ISystemAccountService.cs
@@ -9,8 +9,8 @@ namespace SalesSystem.Application.Interfaces.Services;
 /// </summary>
 public interface ISystemAccountService
 {
-    Task<Result<SystemAccountMappingDto>> GetMappingAsync(SalesSystem.Domain.Accounting.Enums.SystemAccountKey key, int? branchId = null, CancellationToken ct = default);
-    Task<Result<List<SystemAccountMappingDto>>> GetAllMappingsAsync(int? branchId = null, CancellationToken ct = default);
+    Task<Result<SystemAccountMappingDto>> GetMappingAsync(SalesSystem.Domain.Accounting.Enums.SystemAccountKey key, short? branchId = null, CancellationToken ct = default);
+    Task<Result<List<SystemAccountMappingDto>>> GetAllMappingsAsync(short? branchId = null, CancellationToken ct = default);
     Task<Result<SystemAccountMappingDto>> CreateMappingAsync(SalesSystem.Contracts.Requests.CreateSystemAccountMappingRequest request, CancellationToken ct = default);
     Task<Result<SystemAccountMappingDto>> UpdateMappingAsync(int id, SalesSystem.Contracts.Requests.UpdateSystemAccountMappingRequest request, CancellationToken ct = default);
     Task<Result> DeleteMappingAsync(int id, CancellationToken ct = default);
